feat: Update RMQ module registration in API and Auth modules
This commit updates the registration of the RMQ module in the API and Auth modules. It replaces the previous import of `RmqModule` with `RmqModule.register({ name: RmqQueues.Api })` in the `imports` array of both modules. This change ensures that the correct RMQ queue is used for each module.

Note: Please make sure to test the functionality of the updated RMQ module in both the API and Auth services.