feat: Add exclude.util and random.utils to common utils
This commit adds the exclude.util and random.utils files to the common utils directory. The exclude.util file contains a function called "exclude" that allows excluding specific keys from a model object. The random.utils file contains a function called "randomString" that generates a random string of characters. These additions enhance the functionality of the common utils library.

Note: Please make sure to test the functionality of the new files in the common utils library.