feat: Add argon2 and nanoid dependencies
This commit adds the argon2 and nanoid dependencies to the project. The argon2 package is used for secure password hashing, while nanoid is used for generating unique IDs. These dependencies enhance the security and functionality of the project.

Note: Please make sure to test the functionality of the argon2 and nanoid packages.