feat: Add email validation and password hashing to registerAdmin method
This commit adds email validation and password hashing to the registerAdmin method in the AuthService class. The email validation checks if the email ends with "@ahbaco.com" and throws an error if it doesn't. The password is now hashed using the argon2 package before storing it in the database. These enhancements improve the security of the registration process.

Note: Please make sure to test the functionality of the registerAdmin method after these changes.