fix: Update MongoDB URI in local environment configuration
This commit updates the MongoDB URI in the local environment configuration file. The previous URI was set to 'mongodb://mongodb/master', and it has been changed to 'mongodb://mongodb:27017'. This change ensures that the correct MongoDB connection is used in the local environment.

Note: Please make sure to test the functionality of the application with the updated MongoDB URI.